backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
openai==1.57.0
anthropic==0.40.0
supabase==2.10.0
python-multipart==0.0.18
httpx==0.27.2
redis==5.2.1